requirements-dev.txt
# Optional dev dependencies for this framework
ruff>=0.5
jsonschema>=4.0
pyyaml>=6.0
